Add docker, remove emacs code to set PATH
I realize now that doom env is for this specific purpose.
Separately, I now know that mu4e sync is not working because emacs does
not inherit DISPLAY properly, and mbsync passes DISPLAY directly to
pinentry. The result is that pinentry is ran with WAYLAND_DISPLAY
instead of DISPLAY, which does not work.